update to 2.3

+++ src/main/java/org/giiwa/app/web/admin/node.java
@@ -14,16 +14,20 @@
 */
 package org.giiwa.app.web.admin;
 
+import java.util.List;
+
 import javax.servlet.http.HttpServletResponse;
 
 import org.giiwa.bean.*;
-import org.giiwa.conf.Global;
 import org.giiwa.dao.Beans;
 import org.giiwa.dao.X;
 import org.giiwa.dao.Helper.V;
 import org.giiwa.dao.Helper.W;
+import org.giiwa.dao.UID;
 import org.giiwa.json.JSON;
 import org.giiwa.net.mq.MQ;
+import org.giiwa.task.Monitor;
+import org.giiwa.task.Task;
 import org.giiwa.web.*;
 
 /**
@@ -37,6 +41,11 @@ import org.giiwa.web.*;
 public class node extends Controller {
 
 	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
+
+	/**
 	 * Delete.
 	 */
 	@Path(path = "delete", login = true, access = "access.config.admin")
@@ -45,7 +54,12 @@ public class node extends Controller {
 		JSON jo = new JSON();
 
 		String id = this.getString("id");
-		Node.dao.delete(id);
+		Node n = Node.dao.load(id);
+		if (n != null) {
+			Node.dao.delete(id);
+			GLog.oplog.warn("node", "delete", n.label + "/" + n.id, login, node.this.ip());
+		}
+
 		jo.put(X.STATE, 200);
 
 		this.send(jo);
@@ -59,17 +73,22 @@ public class node extends Controller {
 
 		String id = this.getString("id");
 		int power = this.getInt("power");
+		Node n = Node.dao.load(id);
 
-		try {
+		if (n != null) {
+			try {
 
-			Global.setConfig("node." + id, power);
-
-			MQ.topic("giiwa.state",
-					org.giiwa.net.mq.MQ.Request.create().put(JSON.create().append("node", id).append("power", power)));
-		} catch (Exception e) {
-			log.error(e.getMessage(), e);
+				MQ.topic("giiwa.state", org.giiwa.net.mq.MQ.Request.create()
+						.put(JSON.create().append("node", id).append("power", power)));
+				jo.put(X.STATE, 200);
+				jo.put(X.MESSAGE, lang.get("sent.node.power." + power));
+				GLog.oplog.info("node", "power", n.label + ", power=" + power, login, node.this.ip());
+			} catch (Exception e) {
+				GLog.oplog.error("node", "power", e.getMessage(), login, node.this.ip());
+				jo.put(X.STATE, 201);
+				jo.put(X.MESSAGE, e.getMessage());
+			}
 		}
-		jo.put(X.STATE, 200);
 
 		this.send(jo);
 
@@ -81,6 +100,8 @@ public class node extends Controller {
 		String label = this.getString("label");
 		String id = this.getString("id");
 		Node.dao.update(id, V.create("label", label));
+
+		GLog.oplog.error(node.class, "update", label + "/" + id, login, node.this.ip());
 
 		this.send(JSON.create().append(X.STATE, 200));
 
@@ -129,7 +150,7 @@ public class node extends Controller {
 			W q1 = W.create();
 			q1.or("label", name, W.OP.like);
 			q1.or("ip", name, W.OP.like);
-			q1.or("id", name, W.OP.like);
+//			q1.or("id", name, W.OP.like);
 			q.and(q1);
 			this.set("name", name);
 		}
@@ -142,100 +163,76 @@ public class node extends Controller {
 		this.show("/admin/node.index.html");
 	}
 
-	@Path(login = true, path = "usage", access = "access.config.admin")
-	public void usage() {
+	@SuppressWarnings("serial")
+	@Path(path = "add", login = true, access = "access.config.admin")
+	public void add() {
 
-		W q = W.create().sort("label", 1).sort("ip", 1);
+		if (method.isPost()) {
 
-		int s = this.getInt("s");
-		int n = this.getInt("n", 50);
+			String host = this.getString("host");
+			if (X.isEmpty(host)) {
+				this.send(JSON.create().append(X.STATE, 201).append(X.MESSAGE, "ssh missed!"));
+				return;
+			}
 
-		Beans<Node> bs = Node.dao.load(q, s, n);
-		bs.count();
+			String user = this.getString("user");
+			if (X.isEmpty(user)) {
+				this.send(JSON.create().append(X.STATE, 201).append(X.MESSAGE, "user missed!"));
+				return;
+			}
 
-		this.pages(bs, s, n);
+			String passwd = this.getHtml("passwd");
+			if (X.isEmpty(passwd)) {
+				this.send(JSON.create().append(X.STATE, 201).append(X.MESSAGE, "passwd missed!"));
+				return;
+			}
 
-		this.show("/admin/node.usage.html");
-	}
+			String l2 = this.getString("l2");
 
-	@Path(login = true, path = "threads", access = "access.config.admin")
-	public void threads() {
+			String alias = this.getString("alias");
 
-		W q = W.create().sort("label", 1).sort("ip", 1);
+			String access = UID.random(10);
 
-		int s = this.getInt("s");
-		int n = this.getInt("n", 50);
+			long tid = Monitor.start(new Task() {
 
-		Beans<Node> bs = Node.dao.load(q, s, n);
-		bs.count();
+				String message;
+				int state = 0;
 
-		this.pages(bs, s, n);
+				@Override
+				public void onExecute() {
+					Node.add(host, user, passwd, l2, alias, (state, s) -> {
+						this.message = s;
+						this.state = state;
+						Monitor.flush(this);
+					});
+					Monitor.flush(this);
+					if (state == 200) {
+						GLog.oplog.warn(node.class, "add", message, login, node.this.ip());
+					} else {
+						GLog.oplog.error(node.class, "add", message, login, node.this.ip());
+					}
+				}
 
-		this.show("/admin/node.threads.html");
-	}
+			}, access);
 
-	@Path(login = true, path = "tcpestablished", access = "access.config.admin")
-	public void tcpestablished() {
+			this.send(JSON.create().append(X.STATE, 200).append("url", "/f/t/state?id=" + tid + "&access=" + access));
 
-		W q = W.create().sort("label", 1).sort("ip", 1);
+			return;
+		}
 
-		int s = this.getInt("s");
-		int n = this.getInt("n", 50);
+		List<org.giiwa.web.Module> l1 = org.giiwa.web.Module.getAll(true);
+		List<String> l2 = X.asList(l1, e -> {
+			String name = ((org.giiwa.web.Module) e).getName();
+			if (X.isSame(name, "default")) {
+				return null;
+			}
+			return name;
+		});
 
-		Beans<Node> bs = Node.dao.load(q, s, n);
-		bs.count();
+		this.set("l2", X.join(l2, ","));
 
-		this.pages(bs, s, n);
+		this.show("/admin/node.add.html");
 
-		this.show("/admin/node.tcpestablished.html");
-	}
-
-	@Path(login = true, path = "tcpclosewait", access = "access.config.admin")
-	public void tcpclosewait() {
-
-		W q = W.create().sort("label", 1).sort("ip", 1);
-
-		int s = this.getInt("s");
-		int n = this.getInt("n", 50);
-
-		Beans<Node> bs = Node.dao.load(q, s, n);
-		bs.count();
-
-		this.pages(bs, s, n);
-
-		this.show("/admin/node.tcpclosewait.html");
-	}
-
-	@Path(login = true, path = "running", access = "access.config.admin")
-	public void running() {
-
-		W q = W.create().sort("label", 1).sort("ip", 1);
-
-		int s = this.getInt("s");
-		int n = this.getInt("n", 50);
-
-		Beans<Node> bs = Node.dao.load(q, s, n);
-		bs.count();
-
-		this.pages(bs, s, n);
-
-		this.show("/admin/node.running.html");
-	}
-
-	@Path(login = true, path = "pending", access = "access.config.admin")
-	public void pending() {
-
-		W q = W.create().sort("label", 1).sort("ip", 1);
-
-		int s = this.getInt("s");
-		int n = this.getInt("n", 50);
-
-		Beans<Node> bs = Node.dao.load(q, s, n);
-		bs.count();
-
-		this.pages(bs, s, n);
-
-		this.show("/admin/node.pending.html");
 	}
 
 }
